+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 venv
-books.json
+
+.idea
