+++ app/core/config.py
@@ -1,0 +1,1 @@
+ERROR_404 = "Book not found"
