+++ app/dependencies/dependencies.py
@@ -1,4 +1,4 @@
-from src.library_catalog.repository import BookRepository
+from repository import BookRepository
 
 
 async def get_repository():
