+++ requirements.txt
@@ -22,3 +22,4 @@ typing_extensions==4.13.0
 uvicorn==0.34.0
 watchfiles==1.0.4
 websockets==15.0.1
+httpx
