--- src/library_catalog/database.py
@@ -1,3 +1,0 @@
-
-JSON_FILE = "books.json"
-
